Further split out cli configurations (to eventually replace CustomNamespace) and Enum stuff into seperate sources to keep them lightweight and focused (- WIP #316 -)
